+++ src/AceRoutine.h
@@ -31,8 +31,8 @@ SOFTWARE.
 
 #ifndef ACE_ROUTINE_ACE_ROUTINE_H
 
-#include "ace_routine/Routine.h"
-#include "ace_routine/RoutineScheduler.h"
+#include "ace_routine/Coroutine.h"
+#include "ace_routine/CoroutineScheduler.h"
 
 // Version format: xxyyzz == "xx.yy.zz"
 #define ACE_ROUTINE_VERSION 000100